reorganized mesh algorithms and specified CLI
